rename packages: make singular
