+++ traversal/src/main/scala/io/shiftleft/overflowdb/traversals/NodeRefOps.scala
@@ -1,4 +1,4 @@
-package io.shiftleft.overflowdb.traversals
+package io.shiftleft.overflowdb.traversal
 
 import io.shiftleft.overflowdb.NodeRef
 import org.apache.tinkerpop.gremlin.structure.Direction
