+++ traversal/src/main/scala/io/shiftleft/overflowdb/traversals/TraversalSource.scala
@@ -1,4 +1,4 @@
-package io.shiftleft.overflowdb.traversals
+package io.shiftleft.overflowdb.traversal
 
 import io.shiftleft.overflowdb.{NodeRef, OdbGraph}
