+++ traversal/src/test/scala/io/shiftleft/overflowdb/traversal/TraversalTests.scala
@@ -1,4 +1,4 @@
-package io.shiftleft.overflowdb.traversals
+package io.shiftleft.overflowdb.traversal
 
 import org.scalatest.{Matchers, WordSpec}
