+++ traversal/src/test/scala/io/shiftleft/overflowdb/traversal/testdomains/gratefuldead/SongDb.scala
@@ -1,6 +1,6 @@
-package io.shiftleft.overflowdb.traversals.testdomains.gratefuldead
+package io.shiftleft.overflowdb.traversal.testdomains.gratefuldead
 
-import io.shiftleft.overflowdb.traversals.{NodeOps, Traversal}
+import io.shiftleft.overflowdb.traversal.{NodeOps, Traversal}
 import io.shiftleft.overflowdb.{NodeRef, OdbNode, OdbNodeProperty}
 import org.apache.tinkerpop.gremlin.structure.{Direction, VertexProperty}
 import org.apache.tinkerpop.gremlin.util.iterator.IteratorUtils
