+++ traversal/src/test/scala/io/shiftleft/overflowdb/traversal/testdomains/gratefuldead/SungBy.scala
@@ -1,4 +1,4 @@
-package io.shiftleft.overflowdb.traversals.testdomains.gratefuldead
+package io.shiftleft.overflowdb.traversal.testdomains.gratefuldead
 
 import io.shiftleft.overflowdb._
 import scala.jdk.CollectionConverters._
